Stepper functionality finished
